GitHub Actions

+++ .scripts/initSubmodules.ps1
@@ -1,5 +1,6 @@
 $basePath = Join-Path $PSScriptRoot "..\"
 $gitattributesRoot = Join-Path $basePath ".gitattributes"
+$editorConfigRoot = Join-Path $basePath ".editorconfig"
 
 $submodules = Get-Content (Join-Path $basePath ".gitmodules") | Where-Object { $_ -match 'path' } | ForEach-Object { ($_ -split '=' | Select-Object -Last 1 | Out-String).Trim()  }
 
@@ -8,11 +9,10 @@ $submodules | ForEach-Object {
     Write-Output "Processing $submodulePath"
 
     $gitattributesSubmodule = Join-Path $submodulePath ".gitattributes"
+    if (-not (Test-Path $gitattributesSubmodule)) { Copy-Item $gitattributesRoot $gitattributesSubmodule }
 
-    if (-not (Test-Path $gitattributesSubmodule)) {
-        Write-Output "Copying .gitattributes to $submodulePath"
-        Copy-Item $gitattributesRoot $gitattributesSubmodule
-    }
+    $editorConfigSubmodule = Join-Path $submodulePath ".editorconfig"
+    if (-not (Test-Path $editorConfigSubmodule)) { Copy-Item $editorConfigRoot $editorConfigSubmodule }
 
     Push-Location $submodulePath
     git add --renormalize .
